pyproject.toml
[project]
name = "openevent-bot"
version = "1.0.0"
description = "A Telegram bot for creating events in groups with RSVP support"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0,<22.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
openevent-bot = "event_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/event_bot"]
